% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linearregression.R
\name{linearregression}
\alias{linearregression}
\title{Perform linear regression between two variables}
\usage{
linearregression(v1, v2)
}
\arguments{
\item{v1}{A numeric vector}

\item{v2}{A numeric vector}
}
\value{
A linear model object of class \code{lm}
}
\description{
Perform linear regression between two variables
}
\examples{
model <- linearregression(TBQT$age, TBQT$income)
summary(model)
}
