% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TravelBehaviorQinghaiData-package.R
\docType{package}
\name{TravelBehaviorQinghaiData-package}
\alias{TravelBehaviorQinghaiData}
\alias{TravelBehaviorQinghaiData-package}
\title{TravelBehaviorQinghaiData: Data for the paper "Assessing the influence of tourists’ perceived travel environment and their travel behavior on travel satisfaction using structural equation models (SEM): a case of Qinghai-Tibet Plateau"}
\description{
This data was collected between July 21 and August 4 in 2022 through online and on-site questionnaires designed and collected by the Dr. Huaxiong Jiang, Haoran Xu, Kaifei Guo and Lili Zhang. It was designed to collect local and non-local tourists' travel behavior and their perceptions of travel environment on Qinghai-Tibet Plateau. The data was originally compiled in excel format, while in this r package it is processed into a standardized datasets that are reproducible and potentially open-sourced, which meet the objectives of gradudate student course GEOG712 "Reproducible Research Workflow", taught by Prof. Antonio Paez from McMaster University, Canada.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/...}
  \item \url{https://zehuiyin.github.io/...}
  \item Report bugs at \url{https://github.com/zehuiyin/.../issues}
}

}
\author{
\strong{Maintainer}: Haoran Xu \email{xu976@mcmaster.ca} (\href{https://orcid.org/0000-0001-5732-6974}{ORCID})

}
\keyword{internal}
